requirements.txt
termcolor
argparse
dnspython
requests
paramiko